working on styling schema page
